+++ starkbank/corporatepurchase/__init__.py
@@ -1,3 +1,3 @@
 from . import log
 from .log.__log import Log
-from .__corporatepurchase import query, get, parse, response
+from .__corporatepurchase import query, get, page, parse, response